vitraz/requirements.txt
Flask==2.0.1
Pillow==8.3.1
opencv-python==4.5.3.56
numpy==1.21.2
gunicorn==21.2.0